CIV-7302 Update test and latestUpdateSectionBuilder
